Add Dockerfile with fully containerized build
These changes enable a workflow of running a firmware build within a
`docker build` invocation, optionally via a new script at
`scripts/build-docker.sh`, to output build artifacts to `./build/`.
Then, you can run `scripts/flash.sh` on the host (if supported) with
only standard system tools, no build toolchains required.

The Dockerfile is based on Debian "trixie" 13, and uses the Debian
snapshot repositories for consistent package versions, so that later
rebuilds don't silently change what package versions are installed.
Moreover, each Debian package install is preceded by fetching and
cryptographically verifying the package repository contents, ensuring
the repository snapshot is constant through all rebuilds.

This `Dockerfile` deprecates `containers/firmware-open/Containerfile`,
removed by this commit, as it offers several key improvements:

- this Dockerfile actually builds firmware-open, not just sets up the
  toolchains; note it is still possible here to build only the toolchain
  image with just `--target=base`.
- this Dockerfile leverages the standard documented scripts for setting
  up build dependencies, rather than duplicating them in-line, meaning
  less maintenance overhead over time, and better overall testing.
- this Dockerfile builds from local coreboot sources, not remote.
- this Dockerfile being in the repository root enables better
  discoverability and idiomatic interaction with container CLIs, also
  fixing prior hard-coupling to `podman` on the PATH.
- this Dockerfile significantly improves software supply chain security,
  via use of Debian snapshot repositories as described above, and
  preferring toolchains via Debian (e.g. not building own SDCC).

The `docs/building.md` documentation was reorganized to encompass use of
`scripts/build-docker.sh`, akin to prior coverage for host builds.

Given coreboot.org's source archive mirrors serving 403 Forbidden for
weeks, the Dockerfile uses a new flag to stop `scripts/coreboot-sdk.h`
crossgcc build from fetching there, instead just using the default
ftpmirror.gnu.org URLs.

+++ scripts/flash.sh
@@ -31,13 +31,19 @@ BUILD="$(realpath "build/${MODEL}")"
 rm -rf "${BUILD:?}/${BASEDIR}"
 mkdir -p "${BUILD}/${BASEDIR}"
 
-# Rebuild and copy firmware-update
-pushd apps/firmware-update >/dev/null
-    rm -rf "build/x86_64-unknown-uefi"
-    make "build/x86_64-unknown-uefi/boot.efi"
-    cp -v "build/x86_64-unknown-uefi/boot.efi" "${BUILD}/${BASEDIR}"
-    cp -rv "res" "${BUILD}/${BASEDIR}"
-popd >/dev/null
+# Copy firmware-update if pre-built, or otherwise rebuild
+if [ -e "${BUILD}/firmware-update" ]
+then
+    cp -v "${BUILD}/firmware-update/boot.efi" "${BUILD}/${BASEDIR}"
+    cp -rv "${BUILD}/firmware-update/res" "${BUILD}/${BASEDIR}"
+else
+    pushd apps/firmware-update >/dev/null
+        rm -rf "build/x86_64-unknown-uefi"
+        make "build/x86_64-unknown-uefi/boot.efi"
+        cp -v "build/x86_64-unknown-uefi/boot.efi" "${BUILD}/${BASEDIR}"
+        cp -rv "res" "${BUILD}/${BASEDIR}"
+    popd >/dev/null
+fi
 
 # Copy firmware
 mkdir -p "${BUILD}/${BASEDIR}/firmware"
